single-stage auth with c test

+++ tcpcrypt_session.c
@@ -1,6 +1,6 @@
 #include "tcpcrypt_session.h"
 
-int tcpcrypt_get_sid(void)
+unsigned long tcpcrypt_get_sid(void)
 {
     return 1122334455;
 }
