+++ tcpcrypt_session.h
@@ -1,6 +1,6 @@
 #ifndef MOD_AUTH_TCPCRYPT_TCPCRYPT_SESSION_H
 #define MOD_AUTH_TCPCRYPT_TCPCRYPT_SESSION_H
 
-int tcpcrypt_get_sid(void);
+unsigned long tcpcrypt_get_sid(void);
 
 #endif // MOD_AUTH_TCPCRYPT_TCPCRYPT_SESSION_H
